added data files, renamed module
